// chapter7/src/types/movie.ts
export interface Person {
  name: string;
  born?: number;
}


export interface Movie {
  title: string;
  released?: number;
  tagline?: string;
  peopleActedIn?: Person[];
  peopleDirected?: Person[];
}


export interface MovieFormData {
  title: string;
  released?: number;
  tagline?: string;
}

export interface PersonFormData {
  name: string;
  born?: number;
}